Move the context file into its own directory
